organize events, in anticipation of adding more in future
