+++ cibo/events/__init__.py
@@ -1,0 +1,6 @@
+from cibo.events.__event__ import Event
+from cibo.events.connect import Connect
+from cibo.events.disconnect import Disconnect
+from cibo.events.input import Input
+
+_ = Event, Connect, Disconnect, Input
