+++ cibo/events/connect.py
@@ -1,0 +1,15 @@
+from cibo.events import Event
+
+
+class Connect(Event):
+    """Client connection Event."""
+
+    def process(self) -> None:
+        """Process new client connection Events."""
+
+        for client in self._telnet.get_new_clients():
+            client.send_message(
+                "Welcome to cibo.\n"
+                "* Enter 'register name password' to create a new player.\n"
+                "* Enter 'login name password' to log in to an existing player."
+            )
